+++ src/encodings/webp.rs
@@ -107,7 +107,7 @@ fn encode_image<P: Pixel>(
         }
         let sample = image.data[0].color_type();
         if match sample {
-            ColorType::Rgba | ColorType::PaletteRgba | ColorType::LA => {
+            ColorType::Rgba | ColorType::PaletteRgba | ColorType::LumaA => {
                 import_libwebp_picture!(WebPPictureImportRGBA, 4, as_rgba)
             }
             _ => import_libwebp_picture!(WebPPictureImportRGB, 3, as_rgb),