+++ tests/test_webp.rs
@@ -6,7 +6,7 @@ use test_png::COLORS;
 
 #[test]
 fn test_static_webp_encode() -> ril::Result<()> {
-    let image = Image::from_fn(256, 256, |x, _| L(x as u8));
+    let image = Image::from_fn(256, 256, |x, _| Luma(x as u8));
 
     image.save_inferred("tests/out/webp_encode_output.webp")
 }